Restructured thirdparty folder, and added new thirdparty node commands
