+++ Logger.ts
@@ -27,7 +27,7 @@
 
 import * as nslog from 'N/log';
 import * as runtime from 'N/runtime';
-import { type Appender, addAppender, clearAppenders, getLogger, type Logger, logLevel } from './thirdparty/aurelia-logging/aurelia-logging';
+import { type Appender, addAppender, clearAppenders, getLogger, type Logger, logLevel } from './thirdparty/core/aurelia-logging/aurelia-logging';
 
 export {
 	type Appender,
@@ -42,7 +42,7 @@ export {
 	removeAppender,
 	removeCustomLevel,
 	setLevel,
-} from './thirdparty/aurelia-logging/aurelia-logging';
+} from './thirdparty/core/aurelia-logging/aurelia-logging';
 
 /**
  * Value to be prepended to each log message title. Defaults to a random 4 digit integer
@@ -458,6 +458,6 @@ declare var window;
 if (typeof module === 'object') addConsoleAppender(require('aurelia-logging-console'));
 // Else detect NS client script and use NS's async require() to avoid blocking
 else if (typeof console === 'object' && typeof window === 'object' && window.alert)
-	require(['./thirdparty/aurelia-logging/aurelia-logging-console'], addConsoleAppender);
+	require(['./thirdparty/core/aurelia-logging/aurelia-logging-console'], addConsoleAppender);
 // otherwise go ahead and log to the execution log (assume server-side suitescript)
 else addAppender(new ExecutionLogAppender());
